refactor: move state impl classes
