+++ engine/src/main/java/io/camunda/zeebe/streamplatform/state/DbKeyGenerator.java
@@ -5,11 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.processing;
+package io.camunda.zeebe.streamplatform.state;
 
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
-import io.camunda.zeebe.engine.state.NextValueManager;
 import io.camunda.zeebe.engine.state.ZbColumnFamilies;
 import io.camunda.zeebe.protocol.Protocol;
 import io.camunda.zeebe.streamplatform.api.state.KeyGeneratorControls;
