+++ engine/src/main/java/io/camunda/zeebe/streamplatform/state/LastProcessedPosition.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.processing;
+package io.camunda.zeebe.streamplatform.state;
 
 import io.camunda.zeebe.db.DbValue;
 import io.camunda.zeebe.msgpack.UnpackedObject;
