+++ engine/src/main/java/io/camunda/zeebe/streamplatform/state/NextValueManager.java
@@ -5,12 +5,13 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state;
+package io.camunda.zeebe.streamplatform.state;
 
 import io.camunda.zeebe.db.ColumnFamily;
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
 import io.camunda.zeebe.db.impl.DbString;
+import io.camunda.zeebe.engine.state.ZbColumnFamilies;
 import org.agrona.DirectBuffer;
 
 public final class NextValueManager {
