+++ engine/src/test/java/io/camunda/zeebe/engine/state/KeyGeneratorTest.java
@@ -10,10 +10,10 @@ package io.camunda.zeebe.engine.state;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.zeebe.db.ZeebeDb;
-import io.camunda.zeebe.engine.state.processing.DbKeyGenerator;
 import io.camunda.zeebe.engine.util.ZeebeStateRule;
 import io.camunda.zeebe.protocol.Protocol;
 import io.camunda.zeebe.streamplatform.api.state.KeyGenerator;
+import io.camunda.zeebe.streamplatform.state.DbKeyGenerator;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
