+++ engine/src/test/java/io/camunda/zeebe/engine/util/ZeebeStateExtension.java
@@ -17,8 +17,8 @@ import io.camunda.zeebe.engine.state.DefaultZeebeDbFactory;
 import io.camunda.zeebe.engine.state.ZbColumnFamilies;
 import io.camunda.zeebe.engine.state.ZeebeDbState;
 import io.camunda.zeebe.engine.state.mutable.MutableZeebeState;
-import io.camunda.zeebe.engine.state.processing.DbKeyGenerator;
 import io.camunda.zeebe.protocol.Protocol;
+import io.camunda.zeebe.streamplatform.state.DbKeyGenerator;
 import java.io.IOException;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
